moved ai modules into package ai and DB module into package db
