+++ ai/amplifier.py
