+++ ai/emote.py
