+++ ai/gag.py
