+++ ai/identity_enforcer.py
@@ -6,7 +6,7 @@ import discord
 from bot_utils import get_id
 from channels import DRONE_HIVE_CHANNELS
 from roles import DRONE, ENFORCER_DRONE, has_role
-from speech_optimization import informative_status_code_regex, plain_status_code_regex
+from ai.speech_optimization import informative_status_code_regex, plain_status_code_regex
 from webhook import send_webhook
 
 LOGGER = logging.getLogger('ai')
