+++ ai/speech_optimization.py
