+++ ai/stoplights.py
