--- log_test.py
@@ -1,6 +1,0 @@
-import logging
-
-FORMAT = '%(asctime)-15s %(message)s'
-logging.basicConfig(format=FORMAT)
-logger = logging.getLogger('ai')
-logger.warning('drones are good')
